+++ backend/db/dbinit.js
@@ -10,6 +10,8 @@ import {
   assignments,
   assignmentQuestions,
   makePassNullable,
+  addType,
+  addColumns,
 } from '../migrations';
 
 export default async () => {
@@ -30,6 +32,8 @@ export default async () => {
     await db.queryBuilder(assignments);
     await db.queryBuilder(assignmentQuestions);
     await db.queryBuilder(makePassNullable);
+    await db.queryBuilder(addType);
+    await db.queryBuilder(addColumns);
     return true;
   } catch (e) {
     console.log('db init err', e);
