+++ backend/middlewares/authCheck.js
@@ -1,32 +1,95 @@
 import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+import Database from '../db/db';
 
-const Auth = {
-  async verifyToken(request, response, next) {
-    const { token } = request.headers;
-    if (!token) {
-      return response.status(401).send({
-        status: response.statusCode,
-        message: 'Unauthorized, No token provided',
+dotenv.config();
+
+const secretKey = process.env.JWT_PRIVATE_KEY;
+const db = new Database();
+
+exports.verifyToken = async (req, res, next) => {
+  const token = req.headers.authorization.split(' ')[1];
+  if (!token) {
+    return res.status(404).json({
+      status: 'error',
+      error: 'User is not Authenticated',
+    });
+  }
+  try {
+    const verified = await jwt.verify(token, secretKey);
+    req.user = verified;
+    // console.log('req.user');
+    if (!verified) {
+      return res.status(404).json({
+        status: 'error',
+        error: 'Invalid or expired token',
       });
     }
-
-    try {
-      const user = await jwt.verify(token, process.env.JWT_SECRET);
-      if (!user) {
-        return response.status(401).send({
-          status: response.statusCode,
-          message: 'Token expired',
-        });
-      }
-      request.user = { id: user.id };
-      next();
-    } catch (error) {
-      return response.status(401).send({
-        status: response.statusCode,
-        message: 'Invalid token',
-      });
-    }
-  },
+    return next();
+  } catch (err) {
+    res.status(404).json({
+      status: 'error',
+      message: 'Invalid token',
+    });
+  }
 };
 
-export default Auth;
+exports.verifyAuthorUserToken = async (req, res, next) => {
+  const token = req.headers.authorization.split(' ')[1];
+  if (!token) {
+    return res.status(404).json({
+      status: 'error',
+      error: 'User is not Authenticated',
+    });
+  }
+  try {
+    const verified = await jwt.verify(token, secretKey);
+    const userId = verified.id;
+
+    const response = await db.queryBuilder('SELECT * FROM users WHERE id = $1', [userId]);
+    // console.log('response', response.rows);
+    const user = response.rows[0];
+    if (user?.role !== 1) {
+      return res.status(404).json({
+        status: 'error',
+        error: 'Only course owners can perform this action',
+      });
+    }
+    return next();
+  } catch (err) {
+    res.status(404).json({
+      status: 'error',
+      error: err.message,
+    });
+  }
+};
+
+exports.verifyAdminUserToken = async (req, res, next) => {
+  const token = req.headers.authorization.split(' ')[1];
+  if (!token) {
+    return res.status(404).json({
+      status: 'error',
+      error: 'User is not Authenticated',
+    });
+  }
+  try {
+    const verified = await jwt.verify(token, secretKey);
+    const userId = verified.id;
+
+    const response = await db.queryBuilder('SELECT * FROM users WHERE id = $1', [userId]);
+    // console.log('response', response.rows);
+    const user = response.rows[0];
+    if (user?.role !== 2) {
+      return res.status(404).json({
+        status: 'error',
+        error: 'Only admins perform this action',
+      });
+    }
+    return next();
+  } catch (err) {
+    res.status(404).json({
+      status: 'error',
+      error: err.message,
+    });
+  }
+};
